+++ src/handler/inlayHint/buffer.ts
@@ -16,10 +16,16 @@ import workspace from '../../workspace'
 
 export interface InlayHintConfig {
   enable: boolean
+  position: InlayHintPosition,
   display: boolean
   filetypes: string[]
   refreshOnInsertMode: boolean
   enableParameter: boolean
+}
+
+export enum InlayHintPosition {
+  Inline = "inline",
+  Eol = "eol",
 }
 
 let srcId: number | undefined
@@ -70,6 +76,7 @@ export default class InlayHintBuffer implements SyncItem {
     let changed = this._config && this._config.enable != config.enable
     this._config = {
       enable: config.get<boolean>('enable'),
+      position: config.get<InlayHintPosition>('position'),
       display: config.get<boolean>('display', true),
       filetypes: config.get<string[]>('filetypes'),
       refreshOnInsertMode: config.get<boolean>('refreshOnInsertMode'),
@@ -205,7 +212,13 @@ export default class InlayHintBuffer implements SyncItem {
       if (item.paddingRight) {
         chunks.push(nvim.isVim ? [' ', 'Normal'] : [' '])
       }
-      buffer.setVirtualText(srcId, position.line, chunks, { col, hl_mode: 'replace' })
+      if (this.config.position == InlayHintPosition.Eol) {
+        col = 0
+      }
+      // TODO right_gravity field is absent in VirtualTextOption
+      let opts: any = { col, hl_mode: 'replace' }
+      if (!nvim.isVim && item.kind == InlayHintKind.Parameter) { opts.right_gravity = false }
+      buffer.setVirtualText(srcId, position.line, chunks, opts)
     }
     nvim.resumeNotification(true, true)
     this._onDidRefresh.fire()